Removed single updateUser hook to move into a parentHook
* User Update form uses moved hook
* Made an additional check if the user email stayed the same to delete it and not send it to the backend
